sensorhub/sensor_management/pyproject.toml
[tool.poetry]
name = "sensor-management"
version = "0.1.0"
description = "Sensor management service & API"
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
flask-restful = "^0.3.10"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
